updated to pull data from pizzas reducer
